Separate the test files
